把utils的cdTimerSystem.ts內容移動到commands的Timer.ts裡面底下，並改善類別的命名
