+++ src/commands/Timer/clear.ts
@@ -1,10 +1,10 @@
-import { Command } from "@root/src/classes/Command";
-import { Source } from "@root/src/classes/Source";
-import { cancelTimer } from "@root/src/features/utils/cdTimerSystem";
-import { CommandType } from "@root/src/typings/enums";
+import { Command } from "../../classes/Command";
+import { Source } from "../../classes/Source";
+import { CommandType } from "../../typings/enums";
 import { ApplicationCommandOptionType } from "discord.js";
+import { cancelTimer } from "../Timer";
 
-export default class TimerClear extends Command<[Number]> {
+export default class CountdownClear extends Command<[Number]> {
   constructor() {
     super({
       type: CommandType.Utility,
